requirements_ui.txt
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
PyYAML==6.0.1